+++ lib/lectio/student.dart
@@ -10,6 +10,7 @@ import 'package:lectio_wrapper/topics/events/controller.dart';
 import 'package:lectio_wrapper/topics/grades/controller.dart';
 import 'package:lectio_wrapper/topics/gyms/controller.dart';
 import 'package:lectio_wrapper/topics/messages/controller.dart';
+import 'package:lectio_wrapper/topics/students/controller.dart';
 import 'package:lectio_wrapper/topics/weeks/controller.dart';
 import 'package:lectio_wrapper/utils/scraping.dart';
 import 'package:requests/requests.dart';
@@ -37,6 +38,7 @@ class Student {
   late AbsenceController absence;
   late GradeController grades;
   late ContextController context;
+  late StudentsController students;
   Map<String, Uint8List> images = {};
   Student(this.studentId, this.gymId) {
     homework = HomeworkController(this);
@@ -49,6 +51,7 @@ class Student {
     absence = AbsenceController(this);
     grades = GradeController(this);
     context = ContextController(this);
+    students = StudentsController(this);
   }
 
   void setBasicInfo(BasicInfo basicInfo) {