+++ lib/types/context/student.dart
@@ -1,0 +1,6 @@
+import 'package:lectio_wrapper/types/context/team.dart';
+
+class StudentContext extends Context {
+  String imageId;
+  StudentContext(this.imageId);
+}