+++ CHANGELOG.md
@@ -141,3 +141,8 @@
 ### 0.3.2
 
 * Avoid using context for now, as it is very time consuming making long load times.
+
+### 0.3.3
+
+* Get all students as `Person`
+* Fetch student context